Day 9

+++ AdventOfCode2019/Days/Day02/program.cs
@@ -8,7 +8,7 @@ namespace AdventOfCode2019.Days.Day02
 
 		public override object RunPart1()
 		{
-			var code = GetInputString().Split(',').Select(int.Parse).ToArray();
+			var code = GetInputString().Split(',').Select(long.Parse).ToArray();
 			var program = new Intcode(code);
 			foreach (var _ in program.SetValue(1, 12).SetValue(2, 2).Run()) { }
 			return program.GetValue(0);
@@ -16,7 +16,7 @@ namespace AdventOfCode2019.Days.Day02
 
 		public override object RunPart2()
 		{
-			var code = GetInputString().Split(',').Select(int.Parse).ToArray();
+			var code = GetInputString().Split(',').Select(long.Parse).ToArray();
 			for (var noun = 0; noun <= 99; noun++)
 			for (var verb = 0; verb <= 99; verb++)
 			{
